Update error page
