+++ src/components/client/layout/Footer/Footer.styled.tsx
@@ -69,7 +69,7 @@ export const SubscribeLinkWrapper = styled(Grid)(({ theme }) => ({
   },
 
   '&:hover': {
-    'text-decoration': 'underline',
+    textDecoration: 'underline',
     transform: 'scale(1.03)',
     cursor: 'pointer',
     transition: 'all 0.3s ease',
