+++ src/pages/404.tsx
@@ -1,7 +1,7 @@
 import { GetStaticProps } from 'next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 
-import NotFoundPage from 'components/common/errors/NotFoundPage'
+import NotFoundPage from 'components/common/errors/NotFoundPage/NotFoundPage'
 
 export const getStaticProps: GetStaticProps = async ({ locale }) => ({
   props: {
